E2EDAST/requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
alpaca-trade-api==3.2.0
altair==5.4.1
asgiref==3.8.1
astunparse==1.6.3
attrs==24.2.0
backports.functools-lru-cache==2.0.0
beautifulsoup4==4.13.3
blinker==1.8.2
boto3==1.35.32
botocore==1.35.32
cachetools==5.4.0
certifi==2025.1.31
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.41.0
cloudpickle==3.1.1
contourpy==1.2.1
cryptography==43.0.0
cycler==0.12.1
deprecation==2.1.0
dj-database-url==2.2.0
Django==5.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.4.0
django-storages==1.14.4
easyocr==1.7.1
et-xmlfile==1.1.0
exchange_calendars==4.10
Farama-Notifications==0.0.4
filelock==3.15.4
FinRL==0.3.7
Flask==3.0.3
flatbuffers==25.2.10
fonttools==4.53.1
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2024.6.1
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.1
google-auth==2.32.0
google-cloud-vision==3.7.4
google-pasta==0.2.0
googleapis-common-protos==1.63.2
gplearn==0.4.2
grpcio==1.65.4
grpcio-status==1.65.4
gunicorn==23.0.0
gymnasium==1.0.0
h5py==3.13.0
idna==3.7
imageio==2.34.2
indic-transliteration==2.3.69
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keras==3.9.0
kiwisolver==1.4.5
korean-lunar-calendar==0.3.1
langdetect==1.0.9
lazy_loader==0.4
libclang==18.1.1
llvmlite==0.44.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1.post1
mdurl==0.1.2
ml_dtypes==0.5.1
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.0.3
multidict==6.2.0
multitasking==0.0.11
mysql-connector-python==9.0.0
mysqlclient==2.2.4
namex==0.0.8
narwhals==1.9.0
networkx==3.3
ninja==1.11.1.1
nltk==3.9.1
numba==0.61.0
numpy==2.0.1
openai-whisper @ git+https://github.com/openai/whisper.git@517a43ecd132a2089d85f4ebc044728a71d49f6e
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.14.1
packaging==24.1
pandas==2.2.2
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.4
peewee==3.17.9
pillow==10.4.0
platformdirs==4.3.7
propcache==0.3.0
proto-plus==1.24.0
protobuf==5.27.3
psycopg2-binary==2.9.9
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyclipper==1.3.0.post5
pycparser==2.22
pydeck==0.9.1
Pygments==2.18.0
pyluach==2.2.0
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
pyparsing==3.1.2
PyPDF2==3.0.1
pypdfium2==4.30.0
pytesseract==0.3.10
python-bidi==0.6.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
rich==13.9.1
roman==5.0
rpds-py==0.20.0
rsa==4.9
s3transfer==0.10.2
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.0
setuptools==72.1.0
shapely==2.0.5
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
soundfile==0.13.1
soupsieve==2.6
SQLAlchemy==2.0.39
sqlparse==0.5.1
srt==3.5.3
stable_baselines3==2.5.0
stockstats==0.6.4
streamlit==1.39.0
sympy==1.14.0
ta==0.11.0
tenacity==9.0.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==2.5.0
threadpoolctl==3.6.0
tifffile==2024.7.24
tiktoken==0.8.0
tk==0.1.0
toml==0.10.2
toolz==1.0.0
torch==2.7.1
torchaudio==2.7.1
torchvision==0.19.0
tornado==6.4.1
tqdm==4.66.5
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.1
Unidecode==1.3.8
urllib3==1.26.20
vosk==0.3.44
websocket-client==1.8.0
websockets==10.4
Werkzeug==3.0.3
wheel==0.45.1
wrapt==1.17.2
wrds==3.3.0
XlsxWriter==3.2.0
yarl==1.18.3
yfinance==0.2.54
